REF: move misplaced tests (#54429)
* REF: move misplaced extension tests

* remove unnecesssary override
